refactor: split country
