+++ src/lib/info/TrackInfoPopup.svelte
@@ -9,7 +9,6 @@
     import toast from "svelte-french-toast";
     import tippy from "svelte-tippy";
     import { fade, fly } from "svelte/transition";
-    import { db } from "../../data/db";
     import {
         os,
         popupOpen,
@@ -18,13 +17,9 @@
     } from "../../data/store";
     import { focusTrap } from "../../utils/FocusTrap";
     import "../tippy.css";
-    import Input from "../ui/Input.svelte";
 
     import { convertFileSrc, invoke } from "@tauri-apps/api/core";
-    import {
-        fetchAlbumArt,
-        findCountryByArtist,
-    } from "../data/LibraryEnrichers";
+    import { fetchAlbumArt } from "../data/LibraryEnrichers";
     import ButtonWithIcon from "../ui/ButtonWithIcon.svelte";
     import Icon from "../ui/Icon.svelte";
 
@@ -32,6 +27,7 @@
     import { Buffer } from "buffer";
     import LL from "../../i18n/i18n-svelte";
     import MetadataSection from "./MetadataSection.svelte";
+    import CountrySection from "./CountrySection.svelte";
     // optional
 
     // The artwork for this track(s)
@@ -235,42 +231,6 @@
         if ($rightClickedTrack || $rightClickedTracks[0]) {
             reset();
         }
-    }
-
-    let originCountry =
-        ($rightClickedTrack || $rightClickedTracks[0]).originCountry || "";
-    let originCountryEdited = originCountry;
-    let isFetchingOriginCountry = false;
-    function onOriginCountryUpdated(event) {
-        const country = event.target.value;
-        originCountryEdited = country;
-    }
-
-    async function saveTrack() {
-        ($rightClickedTrack || $rightClickedTracks[0]).originCountry =
-            originCountryEdited;
-
-        // Find all songs with this artist
-        const artistSongs = await db.songs
-            .where("artist")
-            .equals(($rightClickedTrack || $rightClickedTracks[0]).artist)
-            .toArray();
-        artistSongs.forEach((s) => {
-            db.songs.update(s.id, { originCountry: originCountryEdited });
-        });
-    }
-
-    async function fetchFromWikipedia() {
-        originCountryEdited = null;
-        isFetchingOriginCountry = true;
-        const country = await findCountryByArtist(
-            ($rightClickedTrack || $rightClickedTracks[0]).artist,
-        );
-        console.log("country", country);
-        if (country) {
-            originCountryEdited = country;
-        }
-        isFetchingOriginCountry = false;
     }
 
     $: isMultiMode = $rightClickedTracks?.length;
@@ -504,49 +464,7 @@
                 {/if}
             </section>
 
-            <section class="enrichment-section boxed">
-                <h5 class="section-title">
-                    <Icon
-                        icon="iconoir:atom"
-                        size={34}
-                    />{$LL.trackInfo.enrichmentCenter()}
-                </h5>
-                <div class="label">
-                    <h4>{$LL.trackInfo.countryOfOrigin()}</h4>
-                    <div
-                        use:tippy={{
-                            content: $LL.trackInfo.countryOfOriginTooltip(),
-                            placement: "right",
-                        }}
-                    >
-                        <Icon icon="mdi:information" />
-                    </div>
-                </div>
-                <div class="country">
-                    <Input
-                        fullWidth
-                        value={originCountryEdited}
-                        placeholder={isFetchingOriginCountry
-                            ? $LL.trackInfo.fetchingOriginCountry()
-                            : ""}
-                        onChange={onOriginCountryUpdated}
-                    />
-                    <ButtonWithIcon
-                        onClick={saveTrack}
-                        text={$LL.trackInfo.save()}
-                        icon="material-symbols:save-outline"
-                        theme="translucent"
-                        disabled={originCountry === originCountryEdited}
-                    />
-                    <ButtonWithIcon
-                        onClick={fetchFromWikipedia}
-                        isLoading={isFetchingOriginCountry}
-                        text={$LL.trackInfo.fetchFromWikipedia()}
-                        icon="tabler:world-download"
-                        theme="transparent"
-                    />
-                </div>
-            </section>
+            <CountrySection />
         </div>
 
         <!-- svelte-ignore a11y-click-events-have-key-events -->
@@ -1119,53 +1037,6 @@
         }
         .find-art-btn {
             margin-top: 1em;
-        }
-    }
-
-    .enrichment-section {
-        margin-top: 1.5em;
-        border: 1px solid
-            color-mix(in srgb, var(--background) 70%, var(--inverse));
-        border-radius: 5px;
-        padding: 2em 1em 1em 1em;
-        grid-column: 1 / 3;
-        background-color: color-mix(
-            in srgb,
-            var(--overlay-bg) 80%,
-            var(--inverse)
-        );
-        position: relative;
-
-        .label {
-            display: flex;
-            flex-direction: row;
-            gap: 5px;
-            align-items: center;
-            margin: 0 0 5px 0;
-
-            h4 {
-                margin: 0;
-                color: var(--text);
-                text-align: left;
-                text-transform: uppercase;
-                letter-spacing: 0.5px;
-                font-size: 0.9em;
-            }
-
-            p {
-                margin: 0;
-                color: rgb(from var(--text) r g b / 0.768);
-            }
-        }
-        .country {
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-            width: fit-content;
-            gap: 5px;
-            p {
-                margin-right: 1em;
-            }
         }
     }
 </style>
